+++ imports/modules/dummy/dummyService.js
@@ -1,7 +1,7 @@
 import {BaseService} from "../shared/service/baseService.js";
 import {DUMMY_LINKS} from "./enums/links";
-import {linkRepository} from "../link/linkRepository";
-import {linkService} from "../link/linkService";
+import {linkRepository} from "../links/linkRepository";
+import {linkService} from "../links/linkService";
 
 export class DummyService extends BaseService {
   constructor(repository, service, data) {
