+++ server/main.js
@@ -1,8 +1,8 @@
 import './init.js';
 import './modules.js';
 import {DummyService} from "../imports/modules/dummy/dummyService";
-import {linkRepository} from "../imports/modules/link/linkRepository";
-import {linkService} from "../imports/modules/link/linkService";
+import {linkRepository} from "../imports/modules/links/linkRepository";
+import {linkService} from "../imports/modules/links/linkService";
 import {DUMMY_LINKS} from "../imports/modules/dummy/enums/links";
 
 Meteor.startup(async () => {
